% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloworld.R
\name{helloworld}
\alias{helloworld}
\title{Prints Hello, World! to the console.}
\usage{
helloworld()
}
\value{
No value returned
}
\description{
Prints Hello, World! to the console.
}
\examples{
helloworld()
}
